refactor: Category와 MinutesWalk enum으로 변경
