+++ src/constants/enums.ts
@@ -1,0 +1,21 @@
+export enum Category {
+  한식 = '한식',
+  중식 = '중식',
+  일식 = '일식',
+  양식 = '양식',
+  아시안 = '아시안',
+  기타 = '기타',
+}
+
+export enum MinutesWalk {
+  '5분 내' = 5,
+  '10분 내' = 10,
+  '15분 내' = 15,
+  '20분 내' = 20,
+  '30분 내' = 30,
+}
+
+export enum Sort {
+  이름순 = '이름순',
+  거리순 = '거리순',
+}
