--- src/constants/setting.ts
@@ -1,7 +1,0 @@
-const TYPE_SETTING = {
-  category: ['한식', '중식', '일식', '양식', '아시안', '기타'] as const,
-  minutesWalk: [5, 10, 15, 20, 30] as const,
-  sort: ['이름순', '거리순'] as const,
-};
-
-export { TYPE_SETTING };
